[feature][1603364560] adds endpoint for login [feature][160364564] adds jwt_authorization to all my endpoints [feature] enables a user to only delete a question that he posted and update an answer that is he or she posted

+++ app/app.py
@@ -1,7 +1,5 @@
 from datetime import datetime
-import json
-
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from flask_restful import Resource, reqparse
 from models.models import AppDb
 
@@ -13,20 +11,19 @@ DTime = datetime.now()
 
 class QuestionDao(object):
 
-    def __init__(self,title, details):
+    def __init__(self,title, details, author):
         self.title = title
         self.details = details
         self.date = str(DTime.date())
+        self.author = author
 
-    def toJSON(self):
-        return json.dumps(self, default=lambda o: o.__dict__,
-                          sort_keys=True, indent=4)
 
 class AnswerDao(object):
 
-    def __init__(self,answer,id):
+    def __init__(self,answer,id,author):
         self.answer = answer
         self.id = id
+        self.author = author
 
 
 class Questions(Resource):
@@ -43,17 +40,18 @@ class Questions(Resource):
         return questions
     @jwt_required
     def post(self):
-
+        name = get_jwt_identity ()
         args = self.reqparse.parse_args()
         title = args['title']
         details = args['details']
+        author = name
 
         if not title.replace(" ", ""):
             return {"message":"title can not be empty"}, 400
         elif not details.replace(" ", ""):
             return {"message":"details can not be empty"}, 400
 
-        questions = QuestionDao(title = title, details= details)
+        questions = QuestionDao(title = title, details= details, author = author)
         if AppDao.check_question_title_exists(questions.title):
             return {"message": "title already used"}, 202
 
@@ -72,14 +70,16 @@ class Question(Resource):
 
     @jwt_required
     def delete(self, id):
-        questions = AppDao.get_question(id)
-        if questions:
-            AppDao.delete_question(id)
-            return{"message":"successfully deleted"}
+        user = get_jwt_identity()
+        questions = AppDao.get_question_with_answers(id)
+        if questions[0][0]['author'] == user:
+            if questions:
+                AppDao.delete_question(id)
+                return{"message":"successfully deleted"}
 
-        else:
-            return {"message": "question id not existing"}
-
+            else:
+                return {"message": "question id not existing"}
+        return {"message":"you can't delete a question you didn't create"}
 
 class Answers(Resource):
 
@@ -90,11 +90,13 @@ class Answers(Resource):
         super(Answers, self).__init__()
     @jwt_required
     def post(self, id):
+        name = get_jwt_identity()
         args = self.reqparse.parse_args()
         answer = args['answer']
+        author = name
         if not answer.replace(" ", ""):
             return {"message":"can't post an empty answer"}, 400
-        answers = AnswerDao(answer=answer, id=id)
+        answers = AnswerDao(answer=answer, id=id, author = author)
         print(answers)
         if AppDao.check_answer_exists(answers.answer):
             return {"message": "answer already posted"}, 400
@@ -118,18 +120,25 @@ class Answer(Resource):
         super(Answer, self).__init__()
     @jwt_required
     def put(self, id,answer_id):
+        name = get_jwt_identity()
         args = self.reqparse.parse_args()
         answer = args['answer']
         if not answer.replace(" ", ""):
             return {"message":"can't post an empty answer"}, 400
-        answers = AnswerDao(answer=answer, id=id)
+        answers = AnswerDao(answer=answer, id=id, author= name)
         print(answers)
         check = AppDao.get_answers(answer_id)
+        question_author = AppDao.get_question_with_answers(id)
         print(check)
-        if check:
-            result =AppDao.update_answer(answers.answer, answer_id)
-            return result
+        if check[0]['user_name'] == name:
+            if check:
+                result =AppDao.update_answer(answers.answer, answer_id)
+                return result
 
-        return {"answer does not exist"}
+            return {"answer does not exist"}
+        elif question_author[0][0]['author'] == name:
+            result = AppDao.update_preferred(answer_id)
+            return  result
+        return {"message":"you are not authorized to update question"}
